support for 'settings.jsonc' file
